third_party/WebKit/LayoutTests/inspector/initial-modules-load-expected.txt

This test validates initial set of loaded modules.


Running: testInitialLoad
Loaded modules:
    elements

Running: testCreateElementsPanel
Loaded modules:
    elements

Running: testCreateNetworkPanel
Loaded modules:
    elements
    network

Running: testShowSourcesPanel
Loaded modules:
    elements
    network
    sources
    sources-formatter-actions

Running: testOpenUISourceCode
Loaded modules:
    elements
    network
    sources
    codemirror
    sources-formatter-actions
